refactor(types): unify BlockId into types::BlockId(u32) + BlockFingerprint
M0.1 from the 2026-04-15 tiered-kv-cache revision
(docs/projects/tiered-kv-cache.md §6 M0.1). Resolves the three-way
BlockId collision flagged in §3 current-state fact 4 and §5.1 of that
doc, and unblocks M1 (wire RadixCache into the scheduler + merge
TierDirectory into Node).

## What changed

Before this commit, three incompatible `BlockId` types lived in the
tree simultaneously, with no `use` path disambiguating them:

- `kv_tier::BlockId(u64)` — content-hash design stub, `todo!()` derive,
  zero non-test callers
- `prefix_cache::BlockId(u32)` — radix-tree node identifier
- `block_manager::BlockId(u32)` — paged KV pool slot allocator

After this commit there is exactly one:

- `infer/src/types.rs::BlockId(pub u32)` — canonical

The other three locations each re-export it (`pub use crate::types::BlockId;`).
Callers that said `use crate::prefix_cache::BlockId;` or
`use crate::block_manager::{BlockId, BlockManager};` keep working
unchanged — the re-export preserves the `crate::…::BlockId` path.

## Why u32 (instead of the original u64 plan)

Worst-case block count under `page_size=16`, `T0=80 GB` HBM, `T1=1 TB`
host-pinned, DeepSeek-V3 (64 layers × 8 KV heads × 128 head_dim × bf16
≈ 256 KB / block) is roughly 4.5 M blocks — well under 2³² with 2×
headroom. vLLM and SGLang also use 32-bit block ids. Keeping it at
`u32` keeps radix-tree nodes cache-line friendly and avoids the
inevitable collision between a "slot id" type (naturally 32-bit) and a
"content hash" type (wants > 64 bits).

## Content hashing lives in a separate type

`types::BlockFingerprint([u8; 16])` — blake3 truncated to 128 bits — is
the new home for what the old `kv_tier::BlockId(u64)` + `BlockHashCtx`
sketch was trying to be. It is only constructed when a block is
actually persisted (tiered-kv-cache M4) or migrated cross-node (M5);
radix-tree nodes carry `Option<BlockFingerprint>` for that use.

The old `BlockHashCtx` struct is deleted — its content-hash stub had
zero callers and its role is subsumed by the `BlockFingerprint`
construction path documented in §5.1.

## Docs updated in the same commit

- `docs/projects/tiered-kv-cache.md` §6 M0.2 — **marked as no-op**.
  The three `prefix_cache.rs` correctness bugs flagged in §8 items
  10–12 (split_node ref_count inherit, lookup ancestor walk, evict
  iterate orphans) were already fixed by commit
  `7c6a704 fix(prefix_cache): split must inherit ref_count + evict
  must cascade` in the 2026-04-13 work batch. The 2026-04-15 survey
  agent's report was stale. Verified: `cargo test -p infer --lib
  prefix_cache` runs 22 tests green including
  `split_node_inherits_ref_count_from_child`,
  `lookup_bumps_every_block_bearing_node_on_path`, and three
  `evict_cascade_*` tests.
- `docs/projects/tiered-kv-cache.md` §14 Next PR rewritten to reflect
  M0.1 shipped, M0.2 no-op, and that M0.3 is now unblocked because
  the `infer-cuda-kernels` crate extraction has landed.
- `docs/plans/tiered-kv-cache-tasks.md` §8 Concrete next action
  rewritten with the same status update.

## Verification

```
cargo fmt --all -- --check                                         ok
cargo check --workspace --no-default-features --features cpu,no-cuda ok
cargo check -p infer --no-default-features --features cuda,no-cuda  ok
cargo test  --workspace --release --no-default-features --features cpu,no-cuda --lib
  → 275 passed, 0 failed, 5 ignored
grep -rn 'pub struct BlockId' infer/src/
  → infer/src/types.rs:25:pub struct BlockId(pub u32)   (exactly 1 hit)
```

Metal check has pre-existing errors in `backend/metal/{mlx,qwen35,
forward}.rs` from an in-flight parallel edit that is not part of M0.1's
scope. Those files are not staged in this commit.

## Does NOT include

- `infer/src/kv_tier/directory.rs` body changes — only its test helper
  signature was updated to match the new u32 type (1 line: `fn
  sample_desc(id: u32, ...)`). The bulk of `directory.rs` is still
  scheduled for deletion in M1 (project doc §5.2), not M0.1.
- M0.3 page_size lift — unblocked but not in this commit. Will be a
  separate PR targeting `crates/infer-cuda-kernels/src/paged_kv.rs`
  and friends.
- M1 RadixCache scheduler wire — depends on M0.1 landing, which is
  what this commit does.

+++ infer/src/kv_tier/directory.rs
@@ -190,7 +190,7 @@ impl Default for TierDirectory {
 mod tests {
     use super::*;
 
-    fn sample_desc(id: u64, tier: Tier, location: BlockLocation) -> BlockDescriptor {
+    fn sample_desc(id: u32, tier: Tier, location: BlockLocation) -> BlockDescriptor {
         BlockDescriptor {
             id: BlockId(id),
             tier,
